Edit: Refactor NotesRequest
* NotesRequest 리펙토링 및 그에 대한 나머지 코드들 수정

+++ server/src/main/java/com/proejct/ClassActionClaim/domain/Notes.java
@@ -17,7 +17,7 @@ public class Notes extends IdDateTimeEntity {
 
     String title;
     String content;
-    Integer week;
+    Long week;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn
@@ -27,7 +27,7 @@ public class Notes extends IdDateTimeEntity {
     @JoinColumn
     Student student;
 
-    public Notes(String title, String content, Integer week, Lecture lecture, Student student) {
+    public Notes(String title, String content, Long week, Lecture lecture, Student student) {
 
         this.title = title;
         this.content = content;
